pyproject.toml
[project]
name = "azurecc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=1.0.0b7",
    "azure-ai-resources>=1.0.0b9",
    "azure-identity>=1.21.0",
    "azure-mgmt-resource>=22.0.0",
    "azure-mgmt-search>=9.1.0",
    "azure-mgmt-storage>=22.1.1",
    "azure-search-documents>=11.5.2",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.8",
    "openai>=1.65.4",
    "owui-pipelines>=0.1.1",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]
